rename EditorWrapper -> Editor
