+++ src/package.js
@@ -22,6 +22,6 @@
 
 import RichTextReader from './components/RichTextReader.vue'
 import ImageResolver from './services/ImageResolver.js'
-import { IMAGE_RESOLVER } from './components/EditorWrapper.provider.js'
+import { IMAGE_RESOLVER } from './components/Editor.provider.js'
 
 export { RichTextReader, ImageResolver, IMAGE_RESOLVER }
